requirements.txt
asgiref==3.3.4
django-taggit==1.4.0
Django==3.2.4
pip==21.1.2
pytz==2021.1
setuptools==57.0.0
sqlparse==0.4.1
